Remove transaction from Deque.__init__
When initializing a Deque, a transaction was used to extend elements from the
given iterable. The transaction is not used in Index.__init__ or in the
FanoutCache.fromcache API.

Users that want Deque.__init__ to use a transaction as before should use:

    d = Deque()
    with d.transact():
        d.extend(iterable)

The transaction is therefore explicit and consistent with other APIs.